Returning data from db when that data exists

+++ zadanie4_go/models/forecast-query-params.go
@@ -1,6 +1,6 @@
 package models
 
 type ForecastQueryParams struct {
-	Latitude  float32 `query:"latitude"`
-	Longitude float32 `query:"longitude"`
+	Latitude  float64 `query:"latitude"`
+	Longitude float64 `query:"longitude"`
 }
