+++ zadanie4_go/models/outer-models/api-weather.go
@@ -1,7 +1,7 @@
 package outer_models
 
 type ApiWeather struct {
-	Latitude        float32
-	Longitude       float32
+	Latitude        float64
+	Longitude       float64
 	Current_Weather ApiWeatherNested
 }
